Aumentando cobertura de teste car controller
